Introduce doctype 'Request for Quotation' (WIP)
