--- senstech/scripts/purchase_order_tools.py
@@ -1,51 +1,0 @@
-# -*- coding: utf-8 -*-
-#
-# purchase_order_tools.py
-# Functions used by client-side scripts for DocType "Purchase Order"
-#
-# Copyright (C) Senstech AG and contributors, 2022
-#
-import frappe
-from frappe import _
-import json, six
-
-
-@frappe.whitelist()
-def transfer_item_drawings(po, items):
-    if isinstance(items, six.string_types):
-        items = json.loads(items)
-    counter = 0
-    for _item in items:
-        item = frappe.get_doc("Item", _item)
-        if item.zeichnung:
-            file_url = item.zeichnung
-            filename = get_file_name(file_url)
-            dt = 'Purchase Order'
-            dn = po
-            if len(check_if_attachment_exists(file_url, dn)) < 1:
-                f = frappe.get_doc({
-                    "doctype": "File",
-                    "file_url": file_url,
-                    "file_name": filename,
-                    "attached_to_doctype": dt,
-                    "attached_to_name": dn,
-                    "folder": 'Home/Attachments',
-                    "file_size": 0,
-                    "is_private": 0
-                })
-                f.flags.ignore_permissions = True
-                try:
-                    f.insert()
-                    frappe.db.commit()
-                    counter += 1
-                except:
-                    pass
-    return counter
-
-
-def get_file_name(file_url):
-    return frappe.db.sql("""SELECT `file_name` FROM `tabFile` WHERE `file_url` = '{file_url}' LIMIT 1""".format(file_url=file_url), as_list=True)[0][0]
-
-
-def check_if_attachment_exists(file_url, dn):
-    return frappe.db.sql("""SELECT `name` FROM `tabFile` WHERE `file_url` = '{file_url}' AND `attached_to_name` = '{dn}'""".format(file_url=file_url, dn=dn), as_list=True)
